removed factory class
