+++ autoscaler/modes/scalecpuandmem.py
@@ -1,13 +1,18 @@
 import operator
 
 from autoscaler.modes.scalemode import AbstractMode
-from autoscaler.modes.modefactory import ModeFactory
+from autoscaler.modes.scalecpu import ScaleByCPU
+from autoscaler.modes.scalemem import ScaleByMemory
 
 
 class ScaleByCPUAndMemory(AbstractMode):
 
     MODE_NAME = 'and'
     MODE_LIST = ['cpu', 'mem']
+    MODES = {
+        'cpu': ScaleByCPU,
+        'mem': ScaleByMemory
+    }
 
     def __init__(self,  api_client=None, app=None, dimension=None):
         super().__init__(api_client, app)
@@ -15,7 +20,7 @@ class ScaleByCPUAndMemory(AbstractMode):
 
         modes = {}
         for idx, mode in enumerate(self.MODE_LIST):
-            modes[mode] = ModeFactory.MODES[mode](
+            modes[mode] = self.MODES[mode](
                 api_client,
                 app,
                 dimension={
